Rename Stubs AB#10444
