+++ Apps/WebClient/src/ClientApp/test/components/navbar.header.test.ts
@@ -7,7 +7,7 @@ import HeaderComponent from "@/components/navmenu/navHeader.vue";
 import { SERVICE_IDENTIFIER } from "@/plugins/inversify";
 import container from "@/plugins/inversify.container";
 import { ILogger } from "@/services/interfaces";
-import { StoreOptionsStub } from "@test/stubs/store/store";
+import { StoreOptionsStub } from "@test/stubs/store/storeOptionsStub";
 
 describe("NavBar Header Component", () => {
     const logger: ILogger = container.get(SERVICE_IDENTIFIER.Logger);
