+++ Apps/WebClient/src/ClientApp/test/router.test.ts
@@ -1,7 +1,15 @@
+import StoreOptionsStub from "@test/stubs/store/storeOptionsStub";
+import { Route } from "vue-router";
+
+import { RegistrationStatus } from "@/constants/registrationStatus";
+import { WebClientConfiguration } from "@/models/configData";
 import { SERVICE_IDENTIFIER, STORE_IDENTIFIER } from "@/plugins/inversify";
 import container from "@/plugins/inversify.container";
 import router, { beforeEachGuard, ClientModule, UserState } from "@/router";
 import { ILogger, IStoreProvider } from "@/services/interfaces";
+import { WinstonLogger } from "@/services/winstonLogger";
+import StoreProvider from "@/store/provider";
+import { GatewayStoreOptions } from "@/store/types";
 import NotFoundComponent from "@/views/errors/notFound.vue";
 import UnauthorizedComponent from "@/views/errors/unauthorized.vue";
 import LandingComponent from "@/views/landing.vue";
@@ -9,13 +17,6 @@ import LoginComponent from "@/views/login.vue";
 import LogoutComponent from "@/views/logout.vue";
 import ProfileComponent from "@/views/profile.vue";
 import RegistrationComponent from "@/views/registration.vue";
-import { Route } from "vue-router";
-import { GatewayStoreOptions } from "@/store/types";
-import { StoreOptionsStub } from "./stubs/store/store";
-import { WebClientConfiguration } from "@/models/configData";
-import { RegistrationStatus } from "@/constants/registrationStatus";
-import { WinstonLogger } from "@/services/winstonLogger";
-import StoreProvider from "@/store/provider";
 
 function flushPromises() {
     return new Promise((resolve) => setImmediate(resolve));
