+++ Apps/WebClient/src/ClientApp/test/stubs/store/authStub.ts
