+++ Apps/WebClient/src/ClientApp/test/stubs/store/configStub.ts
