+++ Apps/WebClient/src/ClientApp/test/stubs/store/encounterStub.ts
