+++ Apps/WebClient/src/ClientApp/test/stubs/store/errorBannerStub.ts
