+++ Apps/WebClient/src/ClientApp/test/stubs/store/idleStub.ts
