+++ Apps/WebClient/src/ClientApp/test/stubs/store/immunizationStub.ts
