+++ Apps/WebClient/src/ClientApp/test/stubs/store/laboratoryStub.ts
