+++ Apps/WebClient/src/ClientApp/test/stubs/store/medicationStub.ts
