+++ Apps/WebClient/src/ClientApp/test/stubs/store/navbarStub.ts
