+++ Apps/WebClient/src/ClientApp/test/stubs/store/noteStub.ts
