+++ Apps/WebClient/src/ClientApp/test/stubs/store/timelineStub.ts
