+++ Apps/WebClient/src/ClientApp/test/stubs/store/userStub.ts
