+++ Apps/WebClient/src/ClientApp/test/timeline.test.ts
@@ -1,24 +1,24 @@
 import "@/plugins/inversify.config";
+
+import StoreOptionsStub from "@test/stubs/store/storeOptionsStub";
 import { createLocalVue, shallowMount, Wrapper } from "@vue/test-utils";
 import VueContentPlaceholders from "vue-content-placeholders";
 import VueRouter from "vue-router";
 import Vuex, { Store } from "vuex";
 
+import { LoadStatus } from "@/models/storeOperations";
 import User from "@/models/user";
 import { SERVICE_IDENTIFIER } from "@/plugins/inversify";
+import container from "@/plugins/inversify.container";
 import { ILogger } from "@/services/interfaces";
+import { ImmunizationState } from "@/store/modules/immunization/types";
 import { GatewayStoreOptions, RootState } from "@/store/types";
 import TimelineView from "@/views/timeline.vue";
 
-import { StoreOptionsStub } from "./stubs/store/store";
-import { ImmunizationState } from "@/store/modules/immunization/types";
-import { LoadStatus } from "@/models/storeOperations";
-import container from "@/plugins/inversify.container";
-
-var store: Store<RootState>;
+let store: Store<RootState>;
 
 function createWrapper(options?: GatewayStoreOptions): Wrapper<TimelineView> {
-    var localVue = createLocalVue();
+    const localVue = createLocalVue();
     localVue.use(Vuex);
     localVue.use(VueRouter);
     localVue.use(VueContentPlaceholders);
@@ -40,20 +40,20 @@ function createWrapper(options?: GatewayStoreOptions): Wrapper<TimelineView> {
 }
 
 describe("Timeline view", () => {
-    var logger: ILogger = container.get(SERVICE_IDENTIFIER.Logger);
+    const logger: ILogger = container.get(SERVICE_IDENTIFIER.Logger);
     logger.initialize("info");
 
     test("is a Vue instance", () => {
-        var wrapper = createWrapper();
+        const wrapper = createWrapper();
         expect(wrapper).toBeTruthy();
     });
 
     test("Loading state", () => {
         // Setup vuex store
-        var options = new StoreOptionsStub();
+        const options = new StoreOptionsStub();
         options.modules.medication.modules.statement.getters.isMedicationStatementLoading = () =>
             true;
-        var wrapper = createWrapper(options);
+        const wrapper = createWrapper(options);
 
         // Check values
         expect(wrapper.find("loadingcomponent-stub").exists()).toBe(true);
@@ -62,10 +62,10 @@ describe("Timeline view", () => {
 
     test("Active", () => {
         // Setup vuex store
-        var options = new StoreOptionsStub();
+        const options = new StoreOptionsStub();
         options.modules.medication.modules.statement.getters.isMedicationStatementLoading = () =>
             false;
-        var wrapper = createWrapper(options);
+        const wrapper = createWrapper(options);
 
         expect(wrapper.find("loadingcomponent-stub").exists()).toBe(false);
         expect(wrapper.find("lineartimeline-stub").isVisible()).toBe(true);
@@ -74,59 +74,59 @@ describe("Timeline view", () => {
 
     test("Shows Calendar", () => {
         // Setup vuex store
-        var options = new StoreOptionsStub();
+        const options = new StoreOptionsStub();
         options.modules.timeline.getters.isLinearView = () => false;
-        var wrapper = createWrapper(options);
+        const wrapper = createWrapper(options);
 
         expect(wrapper.find("lineartimeline-stub").isVisible()).toBe(false);
         expect(wrapper.find("calendartimeline-stub").isVisible()).toBe(true);
     });
 
     test("Shows SMS incomplete profile banner", () => {
-        var user = new User();
+        const user = new User();
         user.hasSMS = true;
         user.verifiedSMS = false;
 
         // Setup vuex store
-        var options = new StoreOptionsStub();
+        const options = new StoreOptionsStub();
         options.modules.user.getters.user = () => user;
-        var wrapper = createWrapper(options);
+        const wrapper = createWrapper(options);
 
         expect(wrapper.find("#incomplete-profile-banner").exists()).toBe(true);
     });
 
     test("Shows Email incomplete profile banner", () => {
-        var user = new User();
+        const user = new User();
         user.hasEmail = true;
         user.verifiedEmail = false;
 
         // Setup vuex store
-        var options = new StoreOptionsStub();
+        const options = new StoreOptionsStub();
         options.modules.user.getters.user = () => user;
-        var wrapper = createWrapper(options);
+        const wrapper = createWrapper(options);
 
         expect(wrapper.find("#incomplete-profile-banner").exists()).toBe(true);
     });
 
     test("Hides incomplete profile banner when verified", () => {
-        var user = new User();
+        const user = new User();
         user.hasEmail = true;
         user.verifiedEmail = true;
         user.hasSMS = true;
         user.verifiedSMS = true;
 
         // Setup vuex store
-        var options = new StoreOptionsStub();
+        const options = new StoreOptionsStub();
         options.modules.user.getters.user = () => user;
-        var wrapper = createWrapper(options);
+        const wrapper = createWrapper(options);
 
         expect(wrapper.find("#incomplete-profile-banner").exists()).toBe(false);
     });
 
     test("Shows Loading immunization", () => {
         // Setup vuex store
-        var options = new StoreOptionsStub();
-        var module = options.modules.immunization;
+        const options = new StoreOptionsStub();
+        const module = options.modules.immunization;
         module.mutations.setStatus = (
             state: ImmunizationState,
             loadStatus: LoadStatus
@@ -138,7 +138,7 @@ describe("Timeline view", () => {
         module.getters.isDeferredLoad = (state: ImmunizationState) =>
             state.status === LoadStatus.DEFERRED;
 
-        var wrapper = createWrapper(options);
+        const wrapper = createWrapper(options);
         expect(wrapper.find("[data-testid=immunizationLoading]").exists()).toBe(
             false
         );
